+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 
 =====================
+21.12.14 - 21.12.14
+=====================
+* Fixed database connection for ADB
+* Added database edition and license model to the CSV extract
+
+=====================
 21.11.16 - 21.11.16
 =====================
 * Added Compute Capacity Reservation